GT-3032: adapt PoC implementation and polish the fidl files

+++ enhanced-position-service/src/position-feedback.cpp
@@ -42,7 +42,7 @@ PositionFeedback::~PositionFeedback()
   return Version;
 }
 
-void PositionFeedback::SetPositionFeedback(const std::vector< std::map< uint16_t, ::DBus::Variant > >& feedback, const uint64_t& timestamp, const uint16_t& feedbackType)
+void PositionFeedback::SetPositionFeedback(const std::vector< std::map< uint64_t, ::DBus::Variant > >& feedback, const uint64_t& timestamp, const uint16_t& feedbackType)
 {
   throw DBus::ErrorNotSupported("Method not supported yet");
 }
